+++ BoardStates.h
@@ -2,6 +2,6 @@
 #include "Definitions.h"
 
 struct BoardStates {
-    int8 shah;
+    int shah;
     BoardStates();
 };
