+++ Debug.h
@@ -5,8 +5,8 @@
 
 class Debug {
 public:
-    static std::vector<Board> GenerateMoves(const Board &position, int8 color);
-    static void printValues(const Board &test_board, int8 color);
+    static std::vector<Board> GenerateMoves(const Board &position, int color);
+    static void printValues(const Board &test_board, int color);
     static void posValuationTest();
     static void boardStatesAndAttackPosTest();
 };
