+++ FigInfo.h
@@ -3,7 +3,7 @@
 
 class FigInfo {
 public:
-    static int8 getPosIndex(int8 fig, int8 color, int8 number = 0);
-    static int8 getFigNumber(int8 pos_index, int8 color);
-    static int8 not(int8 color);
+    static int getPosIndex(int fig, int color, int number = 0);
+    static int getFigNumber(int pos_index, int color);
+    static int not(int color);
 };
