+++ Generator.cpp
@@ -8,7 +8,7 @@
 
 using namespace std;
 
-Generator::Generator(vector<int8> chessboard, vector<int8> positions, vector<int8> colors): m_chessboard(chessboard, positions, colors){
+Generator::Generator(vector<int> chessboard, vector<int> positions, vector<int> colors): m_chessboard(chessboard, positions, colors){
     m_currentFigure[BLACK] = K;          m_currentFigure[WHITE] = K;
     m_currentMove[BLACK] = 0;            m_currentMove[WHITE] = 0;
     m_currentMoveDistance[BLACK] = 1;    m_currentMoveDistance[WHITE] = 1;
@@ -35,7 +35,7 @@ Generator::Generator(const Generator &generator): m_chessboard(generator.m_chess
     m_specialMoves[WHITE] = generator.m_specialMoves[WHITE];
 }
 
-Board Generator::GetNextPosition(int8 color){
+Board Generator::GetNextPosition(int color){
     Board result(m_chessboard);
     if (m_nextPos[color] == -1)
         m_nextPos[color] = nextPos(color);
@@ -55,13 +55,13 @@ Board Generator::GetNextPosition(int8 color){
     return result;
 }
 
-bool Generator::HasNext(int8 color){
+bool Generator::HasNext(int color){
     if (m_nextPos[color] == -1)
         m_nextPos[color] = nextPos(color);
     return (m_nextPos[color] != -1);
 }
 
-vector<Board> Generator::GetAttackMovements(const Board &chessboard, int8 color){
+vector<Board> Generator::GetAttackMovements(const Board &chessboard, int color){
     vector<Board> attack_moves, tmp;
     int figure;
     #pragma omp parallel for default(none) private(figure, tmp) shared(attack_moves, chessboard, color)
@@ -75,18 +75,18 @@ vector<Board> Generator::GetAttackMovements(const Board &chessboard, int8 color)
 
 }
 
-vector<Board> Generator::GetAvailableMovements(const Board &chessboard, int8 color, bool attack_movement){
+vector<Board> Generator::GetAvailableMovements(const Board &chessboard, int color, bool attack_movement){
     vector<Board> normal_moves;
     vector<Board> attack_moves;
     attack_moves = Generator::getAttackPawnMoves(chessboard, color);
     for (int figure : Generator::getFiguresOrder(chessboard, color)){
-        int8 fig_move_idx = FigInfo::getFigNumber(figure, color);
-        int8 curr_pos = chessboard.positions[figure];
-        for (int8 move_idx = 0; move_idx < (int8)g_figMoves[fig_move_idx].size(); move_idx++){
+        int fig_move_idx = FigInfo::getFigNumber(figure, color);
+        int curr_pos = chessboard.positions[figure];
+        for (int move_idx = 0; move_idx < (int)g_figMoves[fig_move_idx].size(); move_idx++){
             if (g_figMoves[fig_move_idx][move_idx] == RANGED_FAR)continue;
             if (*g_figMoves[FigInfo::getFigNumber(figure, color)].rbegin() == RANGED_FAR){
-                for (int8 distance = 1; distance < 8; distance++){
-                    int8 movement = g_figMoves[fig_move_idx][move_idx] * distance;
+                for (int distance = 1; distance < 8; distance++){
+                    int movement = g_figMoves[fig_move_idx][move_idx] * distance;
                     bool atacked = false;
                     if (validateMovement(chessboard, curr_pos, movement, atacked, distance)){
                         if (atacked){
@@ -101,7 +101,7 @@ vector<Board> Generator::GetAvailableMovements(const Board &chessboard, int8 col
                 }
             }
             else{
-                int8 movement = g_figMoves[fig_move_idx][move_idx];
+                int movement = g_figMoves[fig_move_idx][move_idx];
                 bool atacked = false;
                 if (validateMovement(chessboard, curr_pos, movement, atacked, 1)){
                     if (atacked){
@@ -124,12 +124,12 @@ vector<Board> Generator::GetAvailableMovements(const Board &chessboard, int8 col
     return attack_moves;
 }
 
-vector<Board> Generator::GetPosAttackMove(const Board &chessboard, int8 target, int8 color, bool one_attack){
+vector<Board> Generator::GetPosAttackMove(const Board &chessboard, int target, int color, bool one_attack){
     vector<Board> available_moves[3];
-    int8 target_x = target % 8, target_y = target / 8;
-    int8 knights_moves[] = { 1, -2, 2, -1, 2, 1, 1, 2, -1, 2, -2, 1, -2, -1, -1, -2 };
-    int8 distance_moves[] = { 0, -1, 1, -1, 1, 0, 1, 1, 0, 1, -1, 1, -1, 0, -1, -1 };
-    int8 figure;
+    int target_x = target % 8, target_y = target / 8;
+    int knights_moves[] = { 1, -2, 2, -1, 2, 1, 1, 2, -1, 2, -2, 1, -2, -1, -1, -2 };
+    int distance_moves[] = { 0, -1, 1, -1, 1, 0, 1, 1, 0, 1, -1, 1, -1, 0, -1, -1 };
+    int figure;
     if (color == WHITE && target_y != 7){
         if (target_x != 7 && chessboard.board[target + 9] == Pb){
             figure = getFigureFromPosition(chessboard, color, Pb, target + 9);
@@ -192,7 +192,7 @@ vector<Board> Generator::GetPosAttackMove(const Board &chessboard, int8 target, 
     return available_moves[2];
 }
 
-Board Generator::getBoard(const Board &chessboard, int8 color, int8 figure, int8 next_pos){
+Board Generator::getBoard(const Board &chessboard, int color, int figure, int next_pos){
     Board result = chessboard;
     if (result.board[next_pos] != EMPTY)
         Generator::staticDestroyAttackedFig(result, next_pos, color, next_pos);
@@ -204,14 +204,14 @@ Board Generator::getBoard(const Board &chessboard, int8 color, int8 figure, int8
     return result;
 }
 
-bool Generator::pawnCondition(int8 color, int8 next_pos, bool double_move){
+bool Generator::pawnCondition(int color, int next_pos, bool double_move){
     if (double_move){
         if (m_chessboard.board[next_pos - ((-color * 2) - 1) * 8] != EMPTY ||
                 m_chessboard.board[next_pos] != EMPTY)
             return false;
         if (m_chessboard.board[next_pos + 1] == Pb - color ||
                 m_chessboard.board[next_pos - 1] == Pb - color) // beating in passing
-            int8 beating_in_passing = 1;
+            int beating_in_passing = 1;
         return true;
     }
     if (m_currentFigure[color] < 8)//its no pawn
@@ -223,7 +223,7 @@ bool Generator::pawnCondition(int8 color, int8 next_pos, bool double_move){
     return !(m_chessboard.colors[next_pos] != (color ^ 1));
 }
 
-bool Generator::validateMovement(int8 curr_pos, int8 move, bool& attacked){
+bool Generator::validateMovement(int curr_pos, int move, bool& attacked){
     int distance = (m_chessboard.colors[curr_pos] != EMPTY) 
         ? m_currentMoveDistance[m_chessboard.colors[curr_pos]] 
         : 1;
@@ -231,13 +231,13 @@ bool Generator::validateMovement(int8 curr_pos, int8 move, bool& attacked){
 }
 
 //validation of physical opportunity of move no check logic ex. pawn maked 3 steps is validate true
-bool Generator::validateMovement(const Board &chessboard, int8 curr_pos, int8 move, bool& attacked, int8 distance){
+bool Generator::validateMovement(const Board &chessboard, int curr_pos, int move, bool& attacked, int distance){
     bool local_attack_flag = false;
     if (move == RANGED_FAR)
         return false;
     if (chessboard.board[curr_pos] == EMPTY)
         return false;
-    int8 next_pos = curr_pos + move;
+    int next_pos = curr_pos + move;
     if (next_pos < 0 || next_pos > 63)
         return false;
     if (chessboard.board[next_pos] != EMPTY){
@@ -246,7 +246,7 @@ bool Generator::validateMovement(const Board &chessboard, int8 curr_pos, int8 mo
         else
             return false;
     }
-    int8 direction = (move / distance) % 8;
+    int direction = (move / distance) % 8;
     if (direction < 0)direction += 8;
     if (direction > 5){//movement to the left
         if (curr_pos % 8 < (move + curr_pos) % 8)
@@ -261,8 +261,8 @@ bool Generator::validateMovement(const Board &chessboard, int8 curr_pos, int8 mo
     return true;
 }
 
-int8 Generator::getFigureFromPosition(const Board &chessboard, int8 color, int8 figure, int8 position){
-    int8 figure_pos;
+int Generator::getFigureFromPosition(const Board &chessboard, int color, int figure, int position){
+    int figure_pos;
     for (int number = 0; number < 8; number++){
         figure_pos = FigInfo::getPosIndex(figure, color, number);
         if (chessboard.positions[figure_pos] == position)
@@ -272,12 +272,12 @@ int8 Generator::getFigureFromPosition(const Board &chessboard, int8 color, int8 
     return figure_pos;
 }
 
-int8 Generator::manageSpecialMoves(int8 color){
+int Generator::manageSpecialMoves(int color){
     manageSpecialMoves_begin: //goto statement
     if (m_specialMoves[color] <= 8){// pawns double move
-        int8 fig = m_specialMoves[color] + 7 + color * 16;
-        int8 curr_pos = m_chessboard.positions[fig];
-        int8 next_move = curr_pos + ((-color * 2) - 1) * 16;
+        int fig = m_specialMoves[color] + 7 + color * 16;
+        int curr_pos = m_chessboard.positions[fig];
+        int next_move = curr_pos + ((-color * 2) - 1) * 16;
         m_specialMoves[color]++;
         if (!(curr_pos / 8 == color * 5 + 1) || !pawnCondition(color, next_move, true))
             goto manageSpecialMoves_begin;
@@ -290,7 +290,7 @@ int8 Generator::manageSpecialMoves(int8 color){
     return -1;
 }
 
-int8 Generator::nextPos(int8 color){
+int Generator::nextPos(int color){
     nextPos_begin: //goto statement
     if (m_currentFigure[color] > 15){
         if (m_specialMoves[color] == 0){
@@ -300,11 +300,11 @@ int8 Generator::nextPos(int8 color){
         else
             return manageSpecialMoves(color);
     }
-    int8 attack_pawns = m_specialMoves[color] * 2;
-    int8 fig = m_currentFigure[color] + color * 16;
-    int8 curr_pos = m_chessboard.positions[fig];
-    int8 fig_move_idx = FigInfo::getFigNumber(fig, color) + attack_pawns;
-    int8 movement = g_figMoves[fig_move_idx][m_currentMove[color]] * m_currentMoveDistance[color];
+    int attack_pawns = m_specialMoves[color] * 2;
+    int fig = m_currentFigure[color] + color * 16;
+    int curr_pos = m_chessboard.positions[fig];
+    int fig_move_idx = FigInfo::getFigNumber(fig, color) + attack_pawns;
+    int movement = g_figMoves[fig_move_idx][m_currentMove[color]] * m_currentMoveDistance[color];
     bool attacked = false;
     if (curr_pos == DESTROYED || !validateMovement(curr_pos, movement, attacked) || !pawnCondition(color, curr_pos + movement)){
         slideMovement(fig, fig_move_idx);
@@ -318,8 +318,8 @@ int8 Generator::nextPos(int8 color){
     return curr_pos + movement;
 }
 
-int8 Generator::searchAttackFigure(const Board &chessboard, int8 next_pos, int8 color, int move_nb){
-    int8 figure = chessboard.board[next_pos];
+int Generator::searchAttackFigure(const Board &chessboard, int next_pos, int color, int move_nb){
+    int figure = chessboard.board[next_pos];
     switch (figure){
     case H:
         figure = FigInfo::getPosIndex(figure, color);
@@ -343,12 +343,12 @@ int8 Generator::searchAttackFigure(const Board &chessboard, int8 next_pos, int8 
     return figure;
 }
 
-vector<Board> Generator::getAttackPawnMoves(const Board &chessboard, int8 color){
+vector<Board> Generator::getAttackPawnMoves(const Board &chessboard, int color){
     vector<Board> attack_moves;
-    for (int8 figure = 8; figure < 16; figure++){
-        for (int8 i = 0; i < 2; i++){
-            int8 movement = g_figMoves[7 + color][i];
-            int8 curr_pos = chessboard.positions[figure + 16 * color];
+    for (int figure = 8; figure < 16; figure++){
+        for (int i = 0; i < 2; i++){
+            int movement = g_figMoves[7 + color][i];
+            int curr_pos = chessboard.positions[figure + 16 * color];
             bool atacked = false;
             if (curr_pos != DESTROYED && validateMovement(chessboard, curr_pos, movement, atacked, 1)){
                 if (atacked)
@@ -359,11 +359,11 @@ vector<Board> Generator::getAttackPawnMoves(const Board &chessboard, int8 color)
     return attack_moves;
 }
 
-vector<Board> Generator::getSpecialMoves(const Board &chessboard, int8 color){
+vector<Board> Generator::getSpecialMoves(const Board &chessboard, int color){
     vector<Board> double_moves;
-    for (int8 figure = 8; figure < 16; figure++){
-        int8 curr_pos = chessboard.positions[figure + 16 * color];
-        int8 next_move = curr_pos - ((color * 2) - 1) * 16;
+    for (int figure = 8; figure < 16; figure++){
+        int curr_pos = chessboard.positions[figure + 16 * color];
+        int next_move = curr_pos - ((color * 2) - 1) * 16;
         bool atacked = false;
         if ((curr_pos / 8 == color * 5 + 1) &&
                 chessboard.board[next_move + ((color * 2) - 1) * 8] == EMPTY &&
@@ -374,14 +374,14 @@ vector<Board> Generator::getSpecialMoves(const Board &chessboard, int8 color){
     return double_moves;
 }
 
-vector<int8> Generator::getFiguresOrder(const Board &chessboard, int8 color){
-    vector<int8> pos_order = { 27, 28, 35, 36, 18, 19, 20, 21, 26, 29, 34, 37, 42, 43, 44, 45,
+vector<int> Generator::getFiguresOrder(const Board &chessboard, int color){
+    vector<int> pos_order = { 27, 28, 35, 36, 18, 19, 20, 21, 26, 29, 34, 37, 42, 43, 44, 45,
         9, 10, 11, 12, 13, 14, 17, 22, 25, 30, 33, 38, 41, 46, 49, 50,
         51, 52, 53, 54, 0, 1, 2, 3, 4, 5, 6, 7, 56, 57, 58, 59, 60, 61,
         62, 63, 8, 15, 16, 23, 24, 31, 32, 39, 40, 47, 48, 55 };
-    vector<int8> figures_order;
+    vector<int> figures_order;
     for (unsigned pos = 0; pos < pos_order.size(); pos++){
-        for (int8 j = 16 * color; j < 16 + 16 * color; j++){
+        for (int j = 16 * color; j < 16 + 16 * color; j++){
             if (chessboard.positions[j] == pos_order[pos])
                 figures_order.push_back(j);
         }
@@ -389,17 +389,17 @@ vector<int8> Generator::getFiguresOrder(const Board &chessboard, int8 color){
     return figures_order;
 }
 
-void Generator::destroyAttackedFig(Board &board, int8 pos, int8 color, int8 nextPos){
+void Generator::destroyAttackedFig(Board &board, int pos, int color, int nextPos){
     if (nextPos == -1){//sequential mode
         if (m_currentMoveDistance[color] != 1){
-            int8 fig = m_currentFigure[color] + color * 16;
+            int fig = m_currentFigure[color] + color * 16;
             slideMovement(fig, FigInfo::getFigNumber(fig, color));
         }
         m_attack[color] = false;
         nextPos = m_nextPos[color];
     }
-    int8 attacked_fig = FigInfo::getPosIndex(board.board[nextPos], (color + 1) % 2);
-    int8 number = 1;
+    int attacked_fig = FigInfo::getPosIndex(board.board[nextPos], (color + 1) % 2);
+    int number = 1;
     while (board.positions[attacked_fig] != nextPos && number <= 8){
         attacked_fig = FigInfo::getPosIndex(board.board[nextPos], (color + 1) % 2, number);
         number++;
@@ -411,11 +411,11 @@ void Generator::destroyAttackedFig(Board &board, int8 pos, int8 color, int8 next
     board.positions[attacked_fig] = DESTROYED;
 }
 
-void Generator::staticDestroyAttackedFig(Board &board, int8 pos, int8 color, int8 next_pos){
-    int8 attacked_fig = FigInfo::getPosIndex(board.board[next_pos], FigInfo::not(color));
+void Generator::staticDestroyAttackedFig(Board &board, int pos, int color, int next_pos){
+    int attacked_fig = FigInfo::getPosIndex(board.board[next_pos], FigInfo::not(color));
     if (board.board[next_pos] == K)
         board.states.shah = FigInfo::not(color);
-    int8 number = 1;
+    int number = 1;
     while (board.positions[attacked_fig] != next_pos && number < 8){
         attacked_fig = FigInfo::getPosIndex(board.board[next_pos], FigInfo::not(color), number);
         number++;
@@ -424,8 +424,8 @@ void Generator::staticDestroyAttackedFig(Board &board, int8 pos, int8 color, int
     board.positions[attacked_fig] = DESTROYED;
 }
 
-void Generator::slideMovement(int8 fig, int8 fig_move_idx, bool increase_range){
-    int8 color = fig / 16;
+void Generator::slideMovement(int fig, int fig_move_idx, bool increase_range){
+    int color = fig / 16;
     if (increase_range && *g_figMoves[FigInfo::getFigNumber(fig,color)].rbegin() == RANGED_FAR && m_currentMoveDistance[color] < 8){
         m_currentMoveDistance[color]++;
         return;
