+++ Globals.h
@@ -4,6 +4,6 @@
 
 extern Board g_baseBoard;
 extern std::vector<Special*> g_specials;
-extern std::vector<int8> *g_figMoves;
+extern std::vector<int> *g_figMoves;
 
 extern const int c_unknownPosition;
